+++ src/Libraries/Microsoft.Extensions.AI.Abstractions/Utilities/AIJsonUtilities.Schema.Create.cs
@@ -427,7 +427,7 @@ public static partial class AIJsonUtilities
                 if (ResolveAttribute<MinLengthAttribute>() is { } minLengthAttribute)
                 {
                     JsonObject obj = ConvertSchemaToObject(ref schema);
-                    if (obj[TypePropertyName] is JsonNode typeNode && typeNode.GetValueKind() is JsonValueKind.String && typeNode.GetValue<string>() is "string")
+                    if (TryGetSchemaType(obj, out string? schemaType, out _) && schemaType is "string")
                     {
                         obj[MinLengthStringPropertyName] ??= minLengthAttribute.Length;
                     }
@@ -440,7 +440,7 @@ public static partial class AIJsonUtilities
                 if (ResolveAttribute<MaxLengthAttribute>() is { } maxLengthAttribute)
                 {
                     JsonObject obj = ConvertSchemaToObject(ref schema);
-                    if (obj[TypePropertyName] is JsonNode typeNode && typeNode.GetValueKind() is JsonValueKind.String && typeNode.GetValue<string>() is "string")
+                    if (TryGetSchemaType(obj, out string? schemaType, out _) && schemaType is "string")
                     {
                         obj[MaxLengthStringPropertyName] ??= maxLengthAttribute.Length;
                     }
@@ -530,7 +530,7 @@ public static partial class AIJsonUtilities
                 {
                     JsonObject obj = ConvertSchemaToObject(ref schema);
 
-                    if (obj[TypePropertyName] is JsonNode typeNode && typeNode.GetValueKind() is JsonValueKind.String && typeNode.GetValue<string>() is "string")
+                    if (TryGetSchemaType(obj, out string? schemaType, out _) && schemaType is "string")
                     {
                         if (lengthAttribute.MinimumLength > 0)
                         {
@@ -629,6 +629,58 @@ public static partial class AIJsonUtilities
                     }
                 }
 #endif
+#if NET || NETFRAMEWORK
+                static bool TryGetSchemaType(JsonObject schema, [NotNullWhen(true)] out string? schemaType, out bool isNullable)
+                {
+                    schemaType = null;
+                    isNullable = false;
+
+                    if (!schema.TryGetPropertyValue(TypePropertyName, out JsonNode? typeNode))
+                    {
+                        return false;
+                    }
+
+                    switch (typeNode?.GetValueKind())
+                    {
+                        case JsonValueKind.String:
+                            schemaType = typeNode.GetValue<string>();
+                            return true;
+
+                        case JsonValueKind.Array:
+                            string? foundSchemaType = null;
+                            foreach (JsonNode? entry in (JsonArray)typeNode)
+                            {
+                                if (entry?.GetValueKind() is not JsonValueKind.String)
+                                {
+                                    return false;
+                                }
+
+                                string entryValue = entry.GetValue<string>();
+                                if (entryValue is "null")
+                                {
+                                    isNullable = true;
+                                    continue;
+                                }
+
+                                if (foundSchemaType is null)
+                                {
+                                    foundSchemaType = entryValue;
+                                }
+                                else if (foundSchemaType != entryValue)
+                                {
+                                    return false;
+                                }
+                            }
+
+                            schemaType = foundSchemaType;
+                            return schemaType is not null;
+
+                        default:
+                            return false;
+                    }
+                }
+#endif
+
                 TAttribute? ResolveAttribute<TAttribute>()
                     where TAttribute : Attribute
                 {