+++ src/Libraries/Microsoft.Extensions.AI/CHANGELOG.md
@@ -2,12 +2,27 @@
 
 ## NOT YET RELEASED
 
+- Updated the EnableSensitiveData properties on OpenTelemetryChatClient/EmbeddingGenerator to respect a OTEL_INSTRUMENTATION_GENAI_CAPTURE_MESSAGE_CONTENT environment variable.
+- Added OpenTelemetryImageGenerator to provide OpenTelemetry instrumentation for IImageGenerator implementations.
+
+## 9.9.0
+
+- Added `FunctionInvokingChatClient` support for non-invocable tools and `TerminateOnUnknownCalls` property.
+- Added support to `FunctionInvokingChatClient` for user approval of function invocations.
+- Updated the Open Telemetry instrumentation to conform to the latest 1.37.0 draft specification of the Semantic Conventions for Generative AI systems.
+- Fixed `GetResponseAsync<T>` to only look at the contents of the last message in the response.
+
+## 9.8.0
+
 - Added `FunctionInvokingChatClient.AdditionalTools` to allow `FunctionInvokingChatClient` to have access to tools not included in `ChatOptions.Tools` but known to the target service via pre-configuration.
+- Added [Experimental] `IChatReducer` and supporting types
 - Updated to accommodate the additions in `Microsoft.Extensions.AI.Abstractions`.
 
 ## 9.7.1
 
+- Added `FunctionInvokingChatClient.FunctionInvoker` to simplify customizing how functions are invoked.
 - Increased the default `FunctionInvokingChatClient.MaximumIterationsPerRequest` value from 10 to 40.
+- Updated the Open Telemetry instrumentation to conform to the latest 1.36.0 draft specification of the Semantic Conventions for Generative AI systems.
 - Updated to accommodate the additions in `Microsoft.Extensions.AI.Abstractions`.
 
 ## 9.7.0